Добавление отдельного модуля для обработки модальных окон
